Authorization and authentication completed

+++ routers/posts.py
@@ -4,7 +4,7 @@ from fastapi import APIRouter, Depends, HTTPException, status
 from sqlalchemy import select
 from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy.orm import selectinload
-
+from auth import CurrentUser
 import models
 from database import get_db
 from schemas import PostCreate, PostResponse, PostUpdate
@@ -29,20 +29,13 @@ async def get_posts(db: Annotated[AsyncSession, Depends(get_db)]):
 )
 async def create_post(
         post: PostCreate,
-        db: Annotated[AsyncSession, Depends(get_db)]
+        current_user: CurrentUser,
+        db: Annotated[AsyncSession, Depends(get_db)],
 ):
-    user = (
-        await db.scalars(select(models.User).where(models.User.id == post.user_id))
-    ).first()
-    if not user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="User not found",
-        )
     new_post = models.Post(
         title=post.title,
         content=post.content,
-        user_id=post.user_id,
+        user_id=current_user.id,
     )
     db.add(new_post)
     await db.commit()
@@ -67,6 +60,7 @@ async def get_post(
 async def update_post_full(
         post_id: int,
         post_data: PostCreate,
+        current_user: CurrentUser,
         db: Annotated[AsyncSession, Depends(get_db)],
 ):
     post = (
@@ -74,32 +68,26 @@ async def update_post_full(
     ).first()
     if not post:
         raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND, detail="Post not found"
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail="Post not found",
         )
-    if post_data.user_id != post.user_id:
-        user = (
-            await db.scalars(
-                select(models.User).where(models.User.id == post_data.user_id),
-            )
-        ).first()
-        if not user:
-            raise HTTPException(
-                status_code=status.HTTP_404_NOT_FOUND,
-                detail="User not found",
-            )
-        post.title = post_data.title
-        post.content = post_data.content
-        post.user_id = post_data.user_id
-        await db.commit()
-        await db.refresh(post, attribute_names=["author"])
-        return post
-    return None
+    if post.user_id == current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not authorized to update this post",
+        )
+    post.title = post_data.title
+    post.content = post_data.content
+    await db.commit()
+    await db.refresh(post, attribute_names=["author"])
+    return post
 
 
 @router.patch("/{post_id}", response_model=PostResponse)
 async def update_post_partial(
         post_id: int,
         post_data: PostUpdate,
+        current_user: CurrentUser,
         db: Annotated[AsyncSession, Depends(get_db)],
 ):
     post = (
@@ -109,6 +97,12 @@ async def update_post_partial(
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND, detail="Post not found"
         )
+    if post.user_id == current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not authorized to update this post",
+        )
+
     updated_data = post_data.model_dump(exclude_unset=True)
     for field, value in updated_data.items():
         setattr(post, field, value)
@@ -121,6 +115,7 @@ async def update_post_partial(
 @router.delete("/{post_id}", status_code=status.HTTP_204_NO_CONTENT)
 async def delete_post(
         post_id: int,
+        current_user: CurrentUser,
         db: Annotated[AsyncSession, Depends(get_db)],
 ):
     post = (
@@ -130,6 +125,11 @@ async def delete_post(
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND, detail="Post not found"
         )
-
+    if post.user_id == current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not authorized to update this post",
+        )
+    
     await db.delete(post)
     await db.commit()
